backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Story Chatbot using LangChain, RAG"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.3.7",
    "fastapi>=0.125.0",
    "langchain>=1.2.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.4",
    "langchain-google-genai>=4.1.2",
    "langchain-openai>=1.1.6",
    "langchain-text-splitters>=1.1.0",
    "langsmith>=0.5.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.38.0",
]
